fix: share_files migration — use Rails 8 defaults (bigint FK matches bigint PK)
The shares table gets bigint primary keys from Rails 8 regardless of
migration version. So share_files.share_id must also be bigint.
Reverted to t.references default which creates matching bigint FK.

+++ db/schema.rb
@@ -142,7 +142,7 @@ ActiveRecord::Schema[8.0].define(version: 2026_02_19_000000) do
   end
 
   create_table "share_files", force: :cascade do |t|
-    t.integer "share_id", null: false
+    t.bigint "share_id", null: false
     t.string "name", null: false
     t.string "path", null: false
     t.string "relative_path", null: false
